# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Arkova Verification API
  description: |
    Privacy-first document verification anchored to Bitcoin.

    ## Authentication
    - **API Key**: Include `X-API-Key` header with your key (get one at /api/v1/keys)
    - **x402 Payment**: Include `X-Payment` header with base64-encoded payment proof
    - Some endpoints accept both methods

    ## Rate Limits
    - Anonymous: 100 req/min per IP
    - API Key: 1,000 req/min per key
    - Batch: 10 req/min
    - AI endpoints: 30 req/min per user

    Headers returned on every response:
    - `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`
    - `Retry-After` (on 429 responses)
  version: "1.0.0"
  contact:
    name: Arkova Engineering
    url: https://arkova.ai
  license:
    name: Proprietary
servers:
  - url: https://arkova-worker-270018525501.us-central1.run.app/api/v1
    description: Production
  - url: http://localhost:3001/api/v1
    description: Local development
security:
  - apiKey: []
  - x402Payment: []
paths:
...
  /verify/batch:
    post:
      operationId: batchVerify
      summary: Batch verify multiple documents
      description: Submit up to 100 public IDs for batch verification. Returns a job ID for polling.
      tags: [Verification]
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_ids]
              properties:
                public_ids:
                  type: array
                  items:
                    type: string
                  maxItems: 100
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                    enum: [pending]
  /verify/entity:
    get:
      operationId: entityVerify
      summary: Search records for an entity
      description: Search across all records for a specific entity (person or organization).
      tags: [Verification]
      parameters:
        - name: entity
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entity verification results
        "402":
          description: Payment required
  /anchor:
    post:
      operationId: anchorDocument
      summary: Submit a document fingerprint for anchoring
      description: Submits a SHA-256 fingerprint to be anchored to Bitcoin via Merkle batch.
      tags: [Anchoring]
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fingerprint]
              properties:
                fingerprint:
                  type: string
                  description: SHA-256 hash of the document
                  pattern: "^[a-fA-F0-9]{64}$"
                credential_type:
                  type: string
                  enum: [DEGREE, LICENSE, CERTIFICATE, TRANSCRIPT, CLE, PROFESSIONAL, BADGE, ATTESTATION, FINANCIAL, LEGAL, INSURANCE, SEC_FILING, PATENT, REGULATION, PUBLICATION, OTHER]
                description:
                  type: string
                  maxLength: 500
      responses:
        "201":
          description: Anchor created
          content:
            application/json:
              schema:
                type: object
                properties:
                  public_id:
                    type: string
                  status:
                    type: string
                    enum: [PENDING]
                  fingerprint:
                    type: string
  /ai/extract:
    post:
      operationId: aiExtract
      summary: Extract metadata from PII-stripped credential text
      description: Uses AI to extract structured metadata fields from credential text. Input must be PII-stripped.
      tags: [AI]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractionRequest"
      responses:
        "200":
          description: Extraction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractionResult"
  /ai/search:
    post:
      operationId: aiSearch
      summary: Semantic search across records
      description: Search records using natural language queries via AI embeddings.
      tags: [AI]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                limit:
                  type: integer
                  default: 10
                  maximum: 50
      responses:
        "200":
          description: Search results
  /nessie/query:
    post:
      operationId: nessieQuery
      summary: Ask Nessie a compliance question
      description: RAG-powered compliance Q&A using regulatory records.
      tags: [Nessie]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                context:
                  type: string
      responses:
        "200":
          description: Nessie response
        "402":
          description: Payment required
  /compliance/check:
    post:
      operationId: complianceCheck
      summary: Check entity against regulatory records
      tags: [Compliance]
      responses:
        "200":
          description: Compliance check result
        "402":
          description: Payment required
  /keys:
    post:
      operationId: createApiKey
      summary: Create a new API key
      description: Requires Supabase JWT authentication.
      tags: [Key Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: API key created (raw key returned only once)
  /usage:
    get:
      operationId: getUsage
      summary: Get API usage statistics
      tags: [Usage]
      security:
        - apiKey: []
      responses:
        "200":
          description: Usage statistics
  /attestations:
    post:
      operationId: createAttestation
      summary: Create a new attestation
      description: |
        Creates a formal attestation linking an attester to a subject with structured claims.
        The attestation is fingerprinted and queued for Bitcoin anchoring.
      tags: [Attestations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttestationRequest"
            example:
              attestation_type: VERIFICATION
              attester_name: "Acme University Registrar"
              attester_type: INSTITUTION
              attester_title: "Office of the Registrar"
              subject_type: credential
              subject_identifier: "ARK-2026-DEG-001"
              claims:
                - claim: "Degree conferred on 2025-06-15"
                  evidence: "Registrar record #R-44821"
                - claim: "GPA verified above 3.5 threshold"
              summary: "Degree conferral verified by issuing institution"
              jurisdiction: "US-CA"
              expires_at: "2027-06-15T00:00:00Z"
      responses:
        "201":
          description: Attestation created and queued for anchoring
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAttestationResponse"
              example:
                public_id: "ATT-2026-VER-00042"
                attestation_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                fingerprint: "b94d27b9934d3e08a52e52d7da7dabfac484efe37a5380ee9088f7ace2efcde9"
                status: "PENDING"
                verify_url: "https://arkova-worker-270018525501.us-central1.run.app/api/v1/attestations/ATT-2026-VER-00042"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                verified: false
                error: "claims must contain at least one entry"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                verified: false
                error: "unauthorized"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer

    get:
      operationId: listAttestations
      summary: List and filter attestations
      description: |
        Returns a paginated list of attestations. Supports filtering by status,
        attestation type, and subject identifier. Public endpoint — no auth required.
      tags: [Attestations]
      security: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, SECURED, REVOKED, EXPIRED]
          description: Filter by attestation status
        - name: attestation_type
          in: query
          schema:
            type: string
            enum: [VERIFICATION, ENDORSEMENT, AUDIT, APPROVAL, COMPLIANCE, SUPPLY_CHAIN, IDENTITY, WITNESS, CUSTOM]
          description: Filter by attestation type
        - name: subject_identifier
          in: query
          schema:
            type: string
          description: Filter by subject public ID
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor from previous response
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        "200":
          description: Paginated list of attestations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttestationListResponse"
              example:
                data:
                  - public_id: "ATT-2026-VER-00042"
                    attestation_type: VERIFICATION
                    status: SECURED
                    attester_name: "Acme University Registrar"
                    subject_type: credential
                    subject_identifier: "ARK-2026-DEG-001"
                    summary: "Degree conferral verified by issuing institution"
                    created_at: "2026-03-28T12:00:00Z"
                    anchored_at: "2026-03-28T12:15:00Z"
                next_cursor: "eyJpZCI6NDJ9"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
...
  /attestations/batch-verify:
    post:
      operationId: batchVerifyAttestations
      summary: Batch verify multiple attestations
      description: |
        Verify up to 100 attestations in a single request. Returns individual
        verification status for each public ID plus an aggregate summary.
      tags: [Attestations]
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchVerifyAttestationsRequest"
            example:
              public_ids:
                - "ATT-2026-VER-00042"
                - "ATT-2026-END-00099"
                - "ATT-2026-AUD-00007"
      responses:
        "200":
          description: Batch verification results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchVerifyAttestationsResponse"
              example:
                results:
                  - public_id: "ATT-2026-VER-00042"
                    status: SECURED
                    attestation_type: VERIFICATION
                    attester_name: "Acme University Registrar"
                    anchored_at: "2026-03-28T12:15:00Z"
                  - public_id: "ATT-2026-END-00099"
                    status: EXPIRED
                    attestation_type: ENDORSEMENT
                    attester_name: "State Bar of California"
                    anchored_at: "2025-01-10T08:00:00Z"
                    expires_at: "2026-01-10T08:00:00Z"
                  - public_id: "ATT-2026-AUD-00007"
                    error: "not_found"
                summary:
                  total: 3
                  verified: 1
                  not_found: 1
                  expired: 1
                  revoked: 0
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                verified: false
                error: "public_ids must contain between 1 and 100 entries"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: HMAC-SHA256 hashed API key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token
    x402Payment:
      type: apiKey
      in: header
      name: X-Payment
      description: Base64-encoded x402 payment proof
  schemas:
...
    ExtractionRequest:
      type: object
      required: [strippedText, credentialType, fingerprint]
      properties:
        strippedText:
          type: string
          description: PII-stripped credential text (client-side stripping applied)
        credentialType:
          type: string
          enum: [DEGREE, LICENSE, CERTIFICATE, TRANSCRIPT, CLE, PROFESSIONAL, BADGE, ATTESTATION, FINANCIAL, LEGAL, INSURANCE, SEC_FILING, PATENT, REGULATION, PUBLICATION, OTHER]
        fingerprint:
          type: string
          description: SHA-256 hash of the original document
        issuerHint:
          type: string
          description: Optional issuer name hint
    ExtractionResult:
      type: object
      required: [fields, confidence, provider]
      properties:
        fields:
          type: object
          description: Extracted metadata fields
          properties:
            credentialType:
              type: string
            issuerName:
              type: string
            issuedDate:
              type: string
              format: date
            expiryDate:
              type: string
              format: date
            fieldOfStudy:
              type: string
            degreeLevel:
              type: string
            licenseNumber:
              type: string
            accreditingBody:
              type: string
            jurisdiction:
              type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        provider:
          type: string
        tokensUsed:
          type: integer
    CreateAttestationRequest:
      type: object
      required: [attestation_type, attester_name, attester_type, subject_type, subject_identifier, claims]
      properties:
        attestation_type:
          type: string
          enum: [VERIFICATION, ENDORSEMENT, AUDIT, APPROVAL, COMPLIANCE, SUPPLY_CHAIN, IDENTITY, WITNESS, CUSTOM]
          description: Category of attestation
        attester_name:
          type: string
          description: Name of the attesting party
        attester_type:
          type: string
          enum: [INSTITUTION, CORPORATION, INDIVIDUAL, REGULATORY, THIRD_PARTY]
          description: Type of attesting entity
        attester_title:
          type: string
          description: Optional title or role of the attester
        subject_type:
          type: string
          enum: [credential, entity, process, asset]
          description: What the attestation is about
        subject_identifier:
          type: string
          description: Public ID of the subject being attested to
        claims:
          type: array
          minItems: 1
          items:
            type: object
            required: [claim]
            properties:
              claim:
                type: string
                description: The assertion being made
              evidence:
                type: string
                description: Optional supporting evidence reference
          description: One or more structured claims
        summary:
          type: string
          description: Optional human-readable summary
        jurisdiction:
          type: string
          description: Optional jurisdiction code (e.g., US-CA). Omitted when null.
        expires_at:
          type: string
          format: date-time
          description: Optional ISO 8601 expiration timestamp
    CreateAttestationResponse:
      type: object
      required: [public_id, attestation_id, fingerprint, status, verify_url]
      properties:
        public_id:
          type: string
          description: Public identifier for the attestation
        attestation_id:
          type: string
          format: uuid
          description: Internal attestation UUID
        fingerprint:
          type: string
          description: SHA-256 fingerprint of the attestation content
        status:
          type: string
          enum: [PENDING]
          description: Initial status (always PENDING)
        verify_url:
          type: string
          format: uri
          description: Direct URL to verify this attestation
...
    AttestationListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            type: object
            required: [public_id, attestation_type, status, attester_name, subject_type, subject_identifier, created_at]
            properties:
              public_id:
                type: string
              attestation_type:
                type: string
                enum: [VERIFICATION, ENDORSEMENT, AUDIT, APPROVAL, COMPLIANCE, SUPPLY_CHAIN, IDENTITY, WITNESS, CUSTOM]
              status:
                type: string
                enum: [PENDING, SECURED, REVOKED, EXPIRED]
              attester_name:
                type: string
              subject_type:
                type: string
                enum: [credential, entity, process, asset]
              subject_identifier:
                type: string
              summary:
                type: string
              created_at:
                type: string
                format: date-time
              anchored_at:
                type: string
                format: date-time
                nullable: true
        next_cursor:
          type: string
          nullable: true
          description: Cursor for next page, null if no more results
...
    BatchVerifyAttestationsRequest:
      type: object
      required: [public_ids]
      properties:
        public_ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
          description: List of attestation public IDs to verify
    BatchVerifyAttestationsResponse:
      type: object
      required: [results, summary]
      properties:
        results:
          type: array
          items:
            type: object
            required: [public_id]
            properties:
              public_id:
                type: string
              status:
                type: string
                enum: [PENDING, SECURED, REVOKED, EXPIRED]
              attestation_type:
                type: string
                enum: [VERIFICATION, ENDORSEMENT, AUDIT, APPROVAL, COMPLIANCE, SUPPLY_CHAIN, IDENTITY, WITNESS, CUSTOM]
              attester_name:
                type: string
              anchored_at:
                type: string
                format: date-time
                nullable: true
              expires_at:
                type: string
                format: date-time
                nullable: true
              error:
                type: string
                description: Present when attestation could not be found
        summary:
          type: object
          required: [total, verified, not_found, expired, revoked]
          properties:
            total:
              type: integer
            verified:
              type: integer
              description: Count of SECURED attestations
            not_found:
              type: integer
            expired:
              type: integer
            revoked:
              type: integer
    ErrorResponse:
      type: object
      properties:
        verified:
          type: boolean
          example: false
        error:
          type: string
...